make a directory for chainercv.utils.image
