+++ chainercv/utils/image/__init__.py
@@ -1,0 +1,2 @@
+from chainercv.utils.image.read_image import read_image  # NOQA
+from chainercv.utils.image.write_image import write_image  # NOQA
